dio-pdf-chatbot-challenge/dio-pdf-chatbot-challenge/requirements.txt
streamlit>=1.44.0
pypdf>=5.4.0
python-dotenv>=1.0.1
numpy>=1.26.4
openai>=1.68.0
sentence-transformers>=3.4.1
faiss-cpu>=1.8.0.post1
reportlab>=4.1.0
pytest>=8.3.5
